requirements.txt
annotated-types==0.5.0
asyncio==3.4.3
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.3
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.3.0
pydantic==2.4.1
pydantic_core==2.10.1
pytest==7.4.2
python-dotenv==1.0.0
replicate==0.13.0
requests==2.31.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.5
